requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-Mail
python-dotenv
cryptography
APScheduler
requests
vk_api
werkzeug
psycopg2-binary
gunicorn
pytz
beautifulsoup4
feedparser